first version
So the experiment kinda failed.

Let me tell you the step I made.
Which I need to test out. Next time.

I validated that the thread I am running in is the same thread
the editor window is running in.

I found behaviour of EditorAPI weird.
It turns into nil whenever I try to run it on my own.

My first hypothesis was that they unload lua state with commands.
But don't do it with events. Seems like it is not true and Lua state
persists.

I found the way to get the lua objects stored on my C side.
And call the functions. This way I was able to open the dialog box.
From my side.

The next step was to command the editor
to do something for me from my side.

The problem I have found is that in order to do something valuable.
You need to run your function inside the context.

I cannot just call.

```lua
local sheet = EditorAPI.getActiveWorksheet()
-- more things
```
And then do something with it. The documentation states that
the only proper way to do it is via context which they pass you in.

```lua
context.doWithDocument(function (document)
  ...
end)
```

My initial idea was to gather context from the time when the window was
created. Basically window.open(context).

The next problem: it realy does work.
But now you need to click on the editor whenever you need an update.
This is unacceptable so I dig further into it.

I found a way to do what I want by the trick with
  - SetFocus(Window);
  - SetFocus(Editor);

But in the end it seems like the editor is fundamentally not suited
for this kind of work.
Somewhere something corrupts the heap when I store the context.
Or doing this weirdness I am doing.

I still need to find out that it really is context stealing that makes
it.